requirements.txt
astroid==2.6.5
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
cymem==2.0.5
elasticsearch[async]==7.13.4
filelock==3.0.12
huggingface-hub==0.0.12
idna==3.2
isort==5.9.2
Jinja2==3.0.1
joblib==1.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.21.1
packaging==21.0
pathy==0.6.0
Pillow==8.3.1
plac==1.3.3
preshed==3.0.5
pydantic==1.8.2
pylint==2.9.5
pyparsing==2.4.7
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.0
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.1.0
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.8
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
transformers==4.8.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
wasabi==0.8.2
wrapt==1.12.1
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.4.0/en_core_sci_sm-0.4.0.tar.gz
jsonlines